relocated and add more tests
