+++ renthub/tests/test_rental_model.py
@@ -1,4 +1,5 @@
 from django.test import TestCase
+from django.contrib.auth.models import User
 from renthub.models import RoomType, Room, Renter, Rental
 
 
@@ -10,7 +11,7 @@ class RentalModelTest(TestCase):
                                                  ideal_for="One person")
 
         self.room = Room.objects.create(room_number=101,
-                                        detail='Test Room',
+                                        detail='Demo Room',
                                         price=99.99,
                                         availability=True,
                                         room_type=self.room_type)
@@ -23,8 +24,12 @@ class RentalModelTest(TestCase):
 
         self.rental = Rental.objects.create(room=self.room,
                                             renter=self.renter,
-                                            rental_fee=99.99,
+                                            rental_fee=self.room.price,
                                             status="wait")
+
+        self.admin = User.objects.create_user(username="admin",
+                                              email="admin@gmail.com",
+                                              password="adminpass")
 
     def test_rental_creation(self):
         """Test that the Rental instance can be created successfully."""
@@ -32,3 +37,14 @@ class RentalModelTest(TestCase):
         self.assertEqual(str(self.rental.renter), "Pong")
         self.assertEqual(self.rental.rental_fee, 99.99)
         self.assertEqual(self.rental.status, "wait")
+
+    def test_renter_can_create_rental(self):
+        """A renter can create a rental."""
+        rental = Rental.objects.create(room=self.room, renter=self.renter, rental_fee=self.room.price)
+        self.assertIsInstance(rental, Rental)
+        self.assertEqual(rental.renter, self.renter)
+
+    def test_non_renter_cannot_create_rental(self):
+        """A non-renter cannot create a rental."""
+        with self.assertRaises(Exception):
+            Rental.objects.create(room=self.room, renter=self.admin, rental_fee=self.room.price)
